Update for latest rust

+++ src/open_simplex.rs
@@ -33,7 +33,7 @@ pub fn open_simplex2<T: Float>(seed: &Seed, point: &::Point2<T>) -> T {
     fn gradient<T: Float>(seed: &Seed, xs_floor: T, ys_floor: T, dx: T, dy: T) -> T {
         let attn = math::cast::<_, T>(2.0_f64) - dx * dx - dy * dy;
         if attn > Float::zero() {
-            let index = seed.get2::<int>([math::cast(xs_floor), math::cast(ys_floor)]);
+            let index = seed.get2::<isize>([math::cast(xs_floor), math::cast(ys_floor)]);
             let vec = gradient::get2::<T>(index);
             math::pow4(attn) * (dx * vec[0] + dy * vec[1])
         } else {
@@ -41,9 +41,9 @@ pub fn open_simplex2<T: Float>(seed: &Seed, point: &::Point2<T>) -> T {
         }
     }
 
-    let zero: T = math::cast(0u);
-    let one: T = math::cast(1u);
-    let two: T = math::cast(2u);
+    let zero: T = math::cast(0);
+    let one: T = math::cast(1);
+    let two: T = math::cast(2);
     let squish_constant: T = math::cast(SQUISH_CONSTANT_2D);
 
     //Place input coordinates onto grid.
@@ -100,7 +100,7 @@ pub fn open_simplex3<T: Float>(seed: &Seed, point: &::Point3<T>) -> T {
     fn gradient<T: Float>(seed: &Seed, xs_floor: T, ys_floor: T, zs_floor: T, dx: T, dy: T, dz: T) -> T {
         let attn = math::cast::<_, T>(2.0_f64) - dx * dx - dy * dy - dz * dz;
         if attn > Float::zero() {
-            let index = seed.get3::<int>([math::cast(xs_floor), math::cast(ys_floor), math::cast(zs_floor)]);
+            let index = seed.get3::<isize>([math::cast(xs_floor), math::cast(ys_floor), math::cast(zs_floor)]);
             let vec = gradient::get3::<T>(index);
             math::pow4(attn) * (dx * vec[0] + dy * vec[1] + dz * vec[2])
         } else {
